% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_to_c.R
\name{f_to_c}
\alias{f_to_c}
\title{Farenheit Converter}
\usage{
f_to_c(temp_F)
}
\arguments{
\item{temp_F}{Numeric input idicating temperature in Fahreneheit}
}
\value{
Numeric indicating temperatur \code{temp_F} converted to Celsius
}
\description{
This function converts temperature from Fahrenheit to Celsius
}
\examples{
f_to_c(32)
f_to_c(212)

}
\keyword{temperature}
