% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{k-Nearest Neighbor Classification}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{a data frame contains covariates without missing values}

\item{cl}{true class value without NAs for the training data}

\item{k_nn}{integer representing number of neighbors}

\item{k_cv}{number of folds considered}
}
\value{
a list with 2 objects:
\describe{
\item{class}{a vector of the predicted class for all observations}
\item{cv_err}{the cross-validation misclassification error}
}
}
\description{
This functions predicts output class using covariates by k-Nearest Neighbor
classification and evaluate the cross-validation misclassification error.
}
\examples{
penguins_data <- (na.omit(my_penguins))[3:6]
cl <- (na.omit(my_penguins))[,1]
my_knn_cv(penguins_data, cl, 1, 5)
}
\keyword{prediction}
