% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{my_lm function}
\usage{
my_lm(formula, data)
}
\arguments{
\item{formula:}{a formula class object}

\item{data:}{input data frame.}
}
\value{
a table with rows for each coefficient
(including the (Intercept)!) and columns for
\code{the Estimate}, \code{Std. Error},
\code{t value}, and \code{Pr(>|t|)}.
}
\description{
This is a function that fits a linear model in R.
}
\examples{
data(mtcars)
 my_lm(mpg ~ hp + wt,mtcars)

}
\keyword{prediction}
