% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t.test.R
\name{my_t.test}
\alias{my_t.test}
\title{myt.test function}
\usage{
my_t.test(x, alternative, mu)
}
\arguments{
\item{x}{A numeric vector of data.}

\item{alternative}{A character string specifying the alternative hypothesis.}

\item{mu}{A number indicating the null hypothesis value of the mean.}
}
\value{
A list with elements,which includes
\code{test_stat}: the numeric test statistic.
\code{df}: the degrees of freedom.
\code{alternative}: the value of the parameter \code{alternative}.
\code{p_val}: the numeric p-value.
}
\description{
This is a function that performs a one sample t-test in R.
}
\examples{
x <- rnorm(10, mean = 0, sd = 1)
  my_t.test(x, alternative = "two.sided" , mu = 5)

}
\keyword{inference}
