% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t_test.R
\name{my_t_test}
\alias{my_t_test}
\title{my_t_test function}
\usage{
my_t_test(x, alternative, mu)
}
\arguments{
\item{x}{A numeric vector of data.}

\item{alternative}{A character string specifying the alternative hypothesis.
This should only accept "two.sided", "less", or "greater".}

\item{mu}{A number indicating the null hypothesis value of the mean.}
}
\value{
A list with elements,which includes
\code{test_stat}: the numeric test statistic.
\code{df}: the degrees of freedom.
\code{alternative}: the value of the parameter \code{alternative}.
\code{p_val}: the numeric p-value.
}
\description{
This is a function that performs a one sample t-test in R.
}
\examples{
x <- rnorm(10, mean = 0, sd = 1)
  my_t_test(x, alternative = "two.sided" , mu = 5)
y <- rnorm(30, mean = 0, sd = 2)
  my_t_test(y, alternative = "greater" , mu = 10)


}
\keyword{inference}
