% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{Nearest Neighbor Cross-Validation Function}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{\itemize{
\item An input data frame
}}

\item{cl}{\itemize{
\item Vector of the true class values of the training data
}}

\item{k_nn}{\itemize{
\item Numeric representing the number of neighbors
}}

\item{k_cv}{\itemize{
\item Numeric representing the number of folds
}}
}
\value{
A list with two objects:
\itemize {
\item class - Vector with the predicted class values of the testing data
\item cv_err - Numeric representing the rate of misclassification
}
}
\description{
This function uses performs k nearest neighbors to predict an output class and cross validation to
calculate the misclassification rate.
}
\examples{
penguins_data <- na.omit(my_penguins)
true_class <- penguins_data$species
penguins_data <- penguins_data \%>\% dplyr::select(-island, -year, -sex, -species)
my_knn_cv(penguins_data, true_class, 3, 5)
my_knn_cv(penguins_data, true_class, 5, 5)
my_knn_cv(penguins_data, true_class, 10, 5)

}
\keyword{prediction}
