% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{KNN function}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{Dataframe input containing information used to predict class
of each observation.}

\item{cl}{Vector input containing the true class of each data point within
the dataframe \code{train}.}

\item{k_nn}{Numeric input indicating number of nearest neighbors to use
within function.}

\item{k_cv}{Numeric input indicating number of folds used to seperate data.}
}
\value{
A list with elements \code{class} which contains predicted values for
all observations and a numeric \code{cv_err} representing the
cross-validation misclassification error.
}
\description{
This function runs a K-Nearest-Neighbor cross validation
}
\examples{
my_knn_cv(my_penguins[,3:6], my_penguins$species, 1, 5)

}
\keyword{prediction}
