% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{k-Nearest Neighbors Cross-Validation}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{A data frame with training data.}

\item{cl}{True class values of the training data.}

\item{k_nn}{A numeric input representing the number of neighbors.}

\item{k_cv}{A numeric input representing the number of folds.}
}
\value{
A list of predicted class for all observations and CV error.
}
\description{
This function predicts the output class and computes the cross-validation
misclassification error.
}
\examples{
\dontrun{
data <- tidyr::drop_na(my_penguins) \%>\% dplyr::select(-island, -sex, -year)
train <- data \%>\% dplyr::select(-species)
species <- data$species
my_knn_cv(train, species, 1, 5)
my_knn_cv(train, species, 5, 5)
}

}
\keyword{prediction}
