% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{Linear Regression Function}
\usage{
my_lm(formula, data)
}
\arguments{
\item{formula}{A class object used to fit models.}

\item{data}{An input data frame.}
}
\value{
A table with four numbers representing the estimate, standard error, t value, and Pr(>|t|)
for each coefficient.
}
\description{
This function runs a linear regression on a set of data.
}
\examples{
my_lm(mpg ~ hp + wt, data = mtcars)

}
\keyword{inference,}
\keyword{prediction}
