% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{Fitting Linear Models}
\usage{
my_lm(formula, data)
}
\arguments{
\item{formula}{A formula input that describes the model to be fitted.}

\item{data}{A data frame input containing the variables in the model.}
}
\value{
A table with columns for Estimate, Std. Error, t value, Pr(>|t|),
and rows for each coefficient.
}
\description{
This function is used to fit linear models.
}
\examples{
my_lm(mpg ~ hp*wt, mtcars)
my_lm(mpg ~ wt, mtcars)

}
\keyword{inference}
\keyword{prediction}
