% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rf_cv.R
\name{my_rf_cv}
\alias{my_rf_cv}
\title{Random Forest Cross-Validation Function}
\usage{
my_rf_cv(k)
}
\arguments{
\item{k}{\itemize{
\item Numeric indicating the number of folds
}}
}
\value{
A numeric indicating the cross-validation error
}
\description{
This function uses performs random forests to predict the body mass of penguins using the bill length, bill depth,
and flipper length and calculates the misclassification error.
}
\examples{
my_rf_cv(5)
my_rf_cv(10)

}
\keyword{prediction}
