% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rf_cv.R
\name{my_rf_cv}
\alias{my_rf_cv}
\title{Random Forest cross-validation function}
\usage{
my_rf_cv(k)
}
\arguments{
\item{k}{Numeric input for number of folds desired}
}
\value{
Numeric representing the cross-validation error
}
\description{
This function does a Random Forest cross validation
}
\examples{
my_rf_cv(5)
my_rf_cv(8)

}
\keyword{prediction}
