% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rf_cv.R
\name{my_rf_cv}
\alias{my_rf_cv}
\title{Random Forest Cross-Validation}
\usage{
my_rf_cv(k)
}
\arguments{
\item{k}{A numeric input representing the number of folds.}
}
\value{
A numeric output representing the cross-validation error.
}
\description{
This function predicts penguins body mass by bill length, bill depth,
and flipper length by using random forest models, and computes
the cross-validation error.
}
\examples{
my_rf_cv(5)
my_rf_cv(10)

}
