% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t.test.R
\name{my_t.test}
\alias{my_t.test}
\title{Hypothesis Test Function}
\usage{
my_t.test(x, alternative, mu)
}
\arguments{
\item{x}{\itemize{
\item Numeric vector of data
}}

\item{alternative}{\itemize{
\item Character that represents what the alternative hypothesis is
}}

\item{mu}{\itemize{
\item Numeric that represents the null hypothesis value for the mean
}}
}
\value{
A list with the following variables:
\itemize {
\item test_stat - A numeric representing the t value
\item df - A numeric representing the degrees of freedom
\item alternative - A character that represents what the alternative hypothesis is
\item p_val - A numeric representing the p value
}
}
\description{
This function runs a one sample t test.
}
\examples{
my_t.test(my_gapminder$lifeExp, "greater", 59)
my_t.test(my_gapminder$lifeExp, "less", 60)

}
\keyword{inference}
