% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t.test.R
\name{my_t.test}
\alias{my_t.test}
\title{T-Test function}
\usage{
my_t.test(x, alternative = c("two.sided", "less", "greater"), mu)
}
\arguments{
\item{x}{A numeric vector of data}

\item{alternative}{A character string specifying alternate hypothesis. Must
take on the value of "two.sided", "less", or "greater"}

\item{mu}{A numeric indicating the null hypothesis value of the mean}
}
\value{
A list with elements \code{test_stat}, a numeric representing the test statistic,
\code{df}, a numeric representing the degree of freedom,
\code{alternative}, a character string indicating the alternative
hypothesis, and \code{p_val}, a numeric containing the p-value
}
\description{
This function runs a t-test
}
\examples{
set.seed(203)
sample.size <- 100
test_data <- rbinom(sample.size, size = 1, prob = 0.59)
my_t.test(test_data, "two.sided", 0.5)
my_t.test(test_data, "less", mu = 0.5)

}
\keyword{inference}
