% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t.test.R
\name{my_t.test}
\alias{my_t.test}
\title{One Sample t-test Function}
\usage{
my_t.test(x, alternative, mu)
}
\arguments{
\item{x}{A numeric vector of data.}

\item{alternative}{A character string specifying the alternative hypothesis.}

\item{mu}{A numeric input indicating the null hypothesis value of the mean.}
}
\value{
A list with elements: the test statistic, the degrees of freedom,
the value of \code{alternative}, and the numeric p-value.
}
\description{
This function performs a one sample t-test.
}
\examples{
my_t.test(c(1,2,5,7), "two.sided", 0)
my_t.test(c(0,1,2,5,7,-2), "greater", 2)

}
